f prog adv fix

+++ functional_programming_advanced/all.c
@@ -2,12 +2,11 @@
 
 bool all(int *array, size_t len, bool (*func)(int))
 {
-    for (int i = 0; i < len; i++)
+    for (size_t i = 0; i < len; i++)
     {
         if (!func(array[i]))
             return false;
     }
     return true;
 }
-
